prom2016_52/requirements.txt
alembic==0.9.1
backports-abc==0.5
certifi==2017.4.17
click==6.7
coverage==4.4
Django==1.11.1
ExifRead==2.1.2
Flask==0.12.1
Flask-Admin==1.5.0
Flask-HTTPAuth==3.2.2
Flask-Migrate==2.0.3
Flask-MySQL==1.4.0
flask-profiler==1.4
Flask-Script==2.0.5
Flask-SQLAlchemy==2.2
Flask-WTF==0.14.2
image==1.5.5
itsdangerous==0.24
Jinja2==2.9.6
Mako==1.0.6
MarkupSafe==1.0
MySQL-python==1.2.5
olefile==0.44
Pillow==4.1.1
PyMySQL==0.7.11
python-editor==1.0.3
pytz==2017.2
singledispatch==3.4.0.3
six==1.10.0
SQLAlchemy==1.1.9
tornado==4.5.1
Werkzeug==0.12.1
WTForms==2.1
WTForms-JSON==0.3.1
